Correção de alguns itens da classe
Atributos foram encapsulados
